All: Improved GetContent
