+++ templates/widget/home/post.tmpl
@@ -20,6 +20,6 @@ src="{{ index $LatestPost.Metadata "thumbnail" }}"></video>
 </a>
 {{ end }}
 
-{{ $LatestPost.GetContent .Config .Channels .Notes }}
+{{ $LatestPost.GetContent }}
 </div>
 </div>
